Refactor: addPhoto.tsx 리팩토링
