+++ src/components/addPhoto/addPicture/AddPicture.tsx
@@ -2,8 +2,7 @@ import React, { useRef, useState } from 'react';
 
 import IconUploadPicture from '@assets/icons/IconUploadPicture';
 import { uuid } from '@supabase/gotrue-js/dist/module/lib/helpers';
-import imageCompression from 'browser-image-compression';
-import heic2any from 'heic2any';
+import { changeFormatAndCompression } from '@utils/changeImageFormat';
 
 interface TypePicture {
   limit: number;
@@ -16,39 +15,10 @@ const AddPicture = ({ setUploadedFiles, limit }: TypePicture) => {
   const [imgSrcList, setImgSrcList] = useState<string[]>([]);
 
   const onFileChange = async (e: React.ChangeEvent<HTMLInputElement>) => {
-    let file = e.target.files?.[0];
-
-    const options = {
-      maxSizeMB: 5,
-      maxWidthOrHeight: 1920,
-    };
+    const file = e.target.files?.[0];
 
     if (file != null) {
-      if (
-        file.type === 'image/heic' ||
-        file.type === 'image/HEIC' ||
-        file.type === 'image/HEIF' ||
-        file.type === 'image/heif'
-      ) {
-        const convertedBlob = await heic2any({
-          blob: file,
-          toType: 'image/jpeg', // JPEG로 변환
-        });
-
-        const convertedFile = new File(
-          [convertedBlob as Blob],
-          file.name.split('.')[0] + '.jpeg',
-          {
-            type: 'image/jpeg',
-          },
-        );
-
-        file = convertedFile;
-      }
-
-      const compressedFile = await imageCompression(file, options);
-
-      const url = URL.createObjectURL(compressedFile);
+      const { compressedFile, url } = await changeFormatAndCompression(file);
 
       setImgSrcList((prev) => [...prev, url]);
       setUploadedFiles((prev) => [...prev, compressedFile]);
